pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "seungwonme",email = "seungwonan.me@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium (>=4.27.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pre-commit (>=4.0.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
